DEP: Cleanup in old and unused stiffness tensor classes and mixins.
